fix int conversion error

+++ Framework/API/src/IndexProperty.cpp
@@ -52,10 +52,10 @@ Indexing::SpectrumIndexSet IndexProperty::getIndices() const {
     return indexInfo.makeIndexSet();
   } else {
     auto res = std::minmax_element(m_value.cbegin(), m_value.cend());
-    auto min = static_cast<size_t>(*res.first);
-    auto max = static_cast<size_t>(*res.second);
+    auto min = *res.first;
+    auto max = *res.second;
 
-    auto isRange = (max - min) == (m_value.size() - 1);
+    auto isRange = (max - min) == static_cast<int>(m_value.size() - 1);
 
     if (isRange) {
       switch (type) {
